@@ -10,10 +10,11 @@ using System.Threading.Tasks;
 namespace MetaverseMax.Database
 {
     // Using Entity Framework Attributes to define tables (other method is API Fluent)
-    [Keyless]       // Available from EF Core 5 
+    //[Keyless]       // Available from EF Core 5 
     [Table("Owner")]
     public class Owner
     {
+        [Key]
         [Column("owner_matic_key")]
         public string? owner_matic_key { get; set; }
 
@@ -21,7 +22,7 @@ namespace MetaverseMax.Database
         public bool tool_active { get; set; }
 
         [Column("public_key")]
-        public string? public_key { get; set; }        
+        public string? public_key { get; set; }
 
         [Column("type")]
         public int? type { get; set; }
@@ -36,7 +37,7 @@ namespace MetaverseMax.Database
         public int? owner_lookup_count { get; set; }
 
         [Column("district_lookup_count")]
-        public int? district_lookup_count { get; set; }        
+        public int? district_lookup_count { get; set; }
 
         [Column("pet_count")]
         public int? pet_count { get; set; }
@@ -50,6 +51,7 @@ namespace MetaverseMax.Database
         [Column("pro_access_renew_code")]
         public string? pro_access_renew_code { get; set; }
 
+
         // Account can update names - this is the last updated name that is not blank.
         [NotMapped]
         [Column("owner_name")]
@@ -58,5 +60,18 @@ namespace MetaverseMax.Database
         [NotMapped]
         [Column("avatar_id")]
         public int? avatar_id { get; set; }
+
+    }
+
+    [Keyless]       // Available from EF Core 5 
+    [Table("OwnerEXT")]
+    public class OwnerEXT : Owner
+    {
+        // Account can update names - this is the last updated name that is not blank.
+        [Column("owner_name")]
+        public new string? owner_name { get; set; }
+
+        [Column("avatar_id")]
+        public new int? avatar_id { get; set; }
     }
 }
