@@ -20,9 +20,8 @@ namespace MetaverseMax.Database
             {
                 // Select type query using LINQ returning a collection of row matching condition - selecting first row.               
                 ownerSummaryDistrictList = _context.ownerSummaryDistrict
-                    .Where(x => x.district_id == districtId && x.update_instance == updateInstance)
-                    .OrderByDescending(x => x.update_instance)
-                    .ThenByDescending(x => x.owned_plots)
+                    .Where(x => x.district_id == districtId && x.update_instance == updateInstance)                    
+                    .OrderByDescending(x => x.owned_plots)
                     .ToList();
             }
             catch (Exception ex)
