@@ -9,6 +9,8 @@ using System.Data.SqlClient;
 using System.IO;
 using System.Linq;
 using System.Net;
+using System.Net.Http.Json;
+using System.Security.AccessControl;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -175,6 +177,35 @@ namespace MetaverseMax.Database
             return returnCode;
         }
 
+        public RETURN_CODE UpdatePlotRank(int tokenId, decimal influenceEfficiency, bool usePrediction, bool saveEvent)
+        {
+            List<Plot> plotList;
+            RETURN_CODE returnCode = RETURN_CODE.ERROR;
+
+            try
+            {
+                plotList = GetPlotbyToken(tokenId);
+
+                for (int index = 0; index < plotList.Count; index++)
+                {
+                    plotList[index].current_influence_rank = influenceEfficiency;
+                    plotList[index].last_run_produce_predict = usePrediction;
+                }
+                returnCode = RETURN_CODE.SUCCESS;
+
+                if (saveEvent)
+                {
+                    _context.SaveChanges();
+                }
+            }
+            catch (Exception ex)
+            {
+                logException(ex, String.Concat("PlotDB::UpdatePlotRank() : Error updating Plot using token_id : ", tokenId));
+            }
+
+            return returnCode;
+        }
+
         public RETURN_CODE UpdateRelatedBuildingPlot(int plotId)
         {
             RETURN_CODE returnCode = RETURN_CODE.ERROR;
@@ -192,124 +223,136 @@ namespace MetaverseMax.Database
 
             return returnCode;
         }
-
+    
         public Plot AddOrUpdatePlot(JObject jsonContent, int posX, int posY, int plotId, bool saveEvent)
         {
             Plot plotMatched = null, returnPlot = null;
             CitizenManage citizen = new(_context, worldType);
+            BuildingManage buildingManage = new(_context, worldType);
             Building building = new();
+            int newInfluence = 0;
 
             try
             {
-                // Only process if MCP service returned plot obj
-                if (jsonContent != null)
+                // Based on the callers passed plotId, either add a new plot or update an existing plot record.
+                if (plotId == 0)
                 {
-                    // Based on the callers passed plotId, either add a new plot or update an existing plot record.
-                    if (plotId == 0)
+                    // Defensive check Plot may already exist - error occured during initial world sync - plot dropped.
+                    plotMatched = GetPlotbyPosXPosY(posX, posY);
+
+                    if (plotMatched == null)
                     {
-                        // Defensive check Plot may already exist - error occured during initial world sync - plot dropped.
-                        plotMatched = GetPlotbyPosXPosY(posX, posY);
-
-                        if (plotMatched == null)
+                        returnPlot = _context.plot.Add(new Plot()
                         {
-                            returnPlot = _context.plot.Add(new Plot()
-                            {
-                                pos_x = posX,
-                                pos_y = posY,
-                                cell_id = jsonContent.Value<int?>("cell_id") ?? 0,
-                                district_id = jsonContent.Value<int?>("region_id") ?? 0,
-                                land_type = jsonContent.Value<int?>("land_type") ?? 0,
+                            update_type = (int)UPDATE_TYPE.FULL,
+                            pos_x = posX,
+                            pos_y = posY,
+                            cell_id = jsonContent.Value<int?>("cell_id") ?? 0,
+                            district_id = jsonContent.Value<int?>("region_id") ?? 0,
+                            land_type = jsonContent.Value<int?>("land_type") ?? 0,
 
-                                last_updated = DateTime.UtcNow,                                
-                                unclaimed_plot = string.IsNullOrEmpty(jsonContent.Value<string>("owner")),
+                            last_updated = DateTime.UtcNow,                                
+                            unclaimed_plot = string.IsNullOrEmpty(jsonContent.Value<string>("owner")),
 
-                                owner_nickname = jsonContent.Value<string>("owner_nickname"),
-                                owner_matic = jsonContent.Value<string>("owner"),
-                                owner_avatar_id = jsonContent.Value<int>("owner_avatar_id"),
+                            #nullable enable
+                            owner_nickname = jsonContent.Value<string?>("owner_nickname") ?? "",
+                            #nullable disable
+                            owner_matic = jsonContent.Value<string>("owner"),
+                            owner_avatar_id = jsonContent.Value<int>("owner_avatar_id"),
 
-                                resources = jsonContent.Value<int?>("resources") ?? 0,
-                                building_id = jsonContent.Value<int?>("building_id") ?? 0,
-                                building_level = jsonContent.Value<int?>("building_level") ?? 0,
-                                building_type_id = jsonContent.Value<int?>("building_type_id") ?? 0,
-                                token_id = jsonContent.Value<int?>("token_id") ?? 0,
+                            resources = jsonContent.Value<int?>("resources") ?? 0,
+                            building_id = jsonContent.Value<int?>("building_id") ?? 0,
+                            building_level = jsonContent.Value<int?>("building_level") ?? 0,
+                            building_type_id = jsonContent.Value<int?>("building_type_id") ?? 0,
+                            token_id = jsonContent.Value<int?>("token_id") ?? 0,
 
-                                on_sale = jsonContent.Value<bool?>("on_sale") ?? false,
-                                current_price = jsonContent.Value<bool?>("on_sale") ?? false ? building.GetSalePrice(jsonContent.Value<JToken>("sale_data"), worldType) : 0,
+                            on_sale = jsonContent.Value<bool?>("on_sale") ?? false,
+                            current_price = jsonContent.Value<bool?>("on_sale") ?? false ? building.GetSalePrice(jsonContent.Value<JToken>("sale_data"), worldType) : 0,
 
-                                for_rent = (jsonContent.Value<int?>("for_rent") ?? 0) > 0 ? building.GetRentPrice(jsonContent.Value<JToken>("rent_info"), worldType) : 0,
-                                rented = jsonContent.Value<string>("renter") != null,
-                                abundance = jsonContent.Value<int?>("abundance") ?? 0,
-                                building_abundance = 0,
-                                condition = jsonContent.Value<int?>("condition") ?? 0,
+                            for_rent = (jsonContent.Value<int?>("for_rent") ?? 0) > 0 ? building.GetRentPrice(jsonContent.Value<JToken>("rent_info"), worldType) : 0,
+                            rented = jsonContent.Value<string>("renter") != null,
+                            abundance = jsonContent.Value<int?>("abundance") ?? 0,
+                            building_abundance = 0,
+                            condition = jsonContent.Value<int?>("condition") ?? 0,
 
-                                influence_info = GetInfluenceInfoTotal(jsonContent.Value<JToken>("influence_info"), jsonContent.Value<Boolean?>("influence_poi_bonus") ?? false, posX, posY, jsonContent.Value<int?>("building_type_id") ?? 0),
-                                current_influence_rank = 0,
-                                influence = jsonContent.Value<int?>("influence") ?? 0,
-                                influence_bonus = jsonContent.Value<int?>("influence_bonus") ?? 0,
-                                influence_poi_bonus = jsonContent.Value<Boolean?>("influence_poi_bonus") ?? false,
-                                production_poi_bonus = jsonContent.Value<decimal?>("production_poi_bonus") ?? 0.0m,
-                                is_perk_activated = jsonContent.Value<Boolean?>("is_perk_activated") ?? false,
-                                app_4_bonus = GetApplicationBonus(4, jsonContent.Value<JArray>("extra_appliances"), posX, posY),
-                                app_5_bonus = GetApplicationBonus(5, jsonContent.Value<JArray>("extra_appliances"), posX, posY),
-                                app_123_bonus = GetApplicationBonus(5, jsonContent.Value<JArray>("extra_appliances"), posX, posY),
+                            influence_info = GetInfluenceInfoTotal(jsonContent.Value<JToken>("influence_info"), jsonContent.Value<Boolean?>("influence_poi_bonus") ?? false, posX, posY, jsonContent.Value<int?>("building_type_id") ?? 0),
+                            current_influence_rank = 0,
+                            influence = jsonContent.Value<int?>("influence") ?? 0,
+                            influence_bonus = jsonContent.Value<int?>("influence_bonus") ?? 0,
+                            influence_poi_bonus = jsonContent.Value<Boolean?>("influence_poi_bonus") ?? false,
+                            production_poi_bonus = jsonContent.Value<decimal?>("production_poi_bonus") ?? 0.0m,
+                            is_perk_activated = jsonContent.Value<Boolean?>("is_perk_activated") ?? false,
+                            app_4_bonus = GetApplicationBonus(4, jsonContent.Value<JArray>("extra_appliances"), posX, posY),
+                            app_5_bonus = GetApplicationBonus(5, jsonContent.Value<JArray>("extra_appliances"), posX, posY),
+                            app_123_bonus = GetApplicationBonus(5, jsonContent.Value<JArray>("extra_appliances"), posX, posY),
 
-                                low_stamina_alert = citizen.CheckCitizenStamina(jsonContent.Value<JArray>("citizens"), jsonContent.Value<int?>("building_type_id") ?? 0),
-                                action_id = jsonContent.Value<int?>("action_id") ?? 0,
+                            citizen_count = jsonContent.Value<JArray>("citizens") == null ? 0 : jsonContent.Value<JArray>("citizens").Count,
+                            low_stamina_alert = citizen.CheckCitizenStamina(jsonContent.Value<JArray>("citizens"), jsonContent.Value<int?>("building_type_id") ?? 0),
+                            action_id = jsonContent.Value<int?>("action_id") ?? 0,
 
-                                predict_produce = 0,
-                                last_run_produce_id = 0,
-                                last_run_produce_predict = false
-                            }).Entity;
-                        }
-                        else
-                        {
-                            _context.LogEvent(String.Concat("PlotDB:AddOrUpdatePlot() : Existing plot was found for X:", posX, " Y:", posY,". This maybe unexpected! Call was to create new plot at these XY coord. Existing Plot will be updated"));
-                        }
+                            predict_produce = 0,
+                            last_run_produce_id = 0,
+                            last_run_produce_predict = false
+
+                        }).Entity;
                     }
-                    
-                    if (plotId != 0 || plotMatched != null)
+                    else
                     {
-                        plotMatched = plotMatched == null ? _context.plot.Find(plotId) : plotMatched;
-
-                        plotMatched.last_updated = DateTime.UtcNow;
-
-                        plotMatched.unclaimed_plot = string.IsNullOrEmpty(jsonContent.Value<string>("owner"));
-                        plotMatched.owner_nickname = jsonContent.Value<string>("owner_nickname");
-                        plotMatched.owner_matic = jsonContent.Value<string>("owner");
-                        plotMatched.owner_avatar_id = jsonContent.Value<int>("owner_avatar_id");
-                        plotMatched.resources = jsonContent.Value<int?>("resources") ?? 0;
-                        plotMatched.building_id = jsonContent.Value<int?>("building_id") ?? 0;
-                        plotMatched.building_level = jsonContent.Value<int?>("building_level") ?? 0;
-                        plotMatched.building_type_id = jsonContent.Value<int?>("building_type_id") ?? 0;
-                        plotMatched.token_id = jsonContent.Value<int?>("token_id") ?? 0;
-
-                        plotMatched.on_sale = jsonContent.Value<bool?>("on_sale") ?? false;
-                        plotMatched.current_price = plotMatched.on_sale ? building.GetSalePrice(jsonContent.Value<JToken>("sale_data"), worldType) : 0;
-                        
-                        plotMatched.for_rent = (jsonContent.Value<int?>("for_rent") ?? 0) > 0 ? building.GetRentPrice(jsonContent.Value<JToken>("rent_info"), worldType) : 0;
-                        plotMatched.rented = jsonContent.Value<string>("renter") != null;
-                        plotMatched.abundance = jsonContent.Value<int?>("abundance") ?? 0;
-                        plotMatched.building_abundance = plotMatched.building_abundance;        // only update from BuildingManage class calc
-                        plotMatched.condition = jsonContent.Value<int?>("condition") ?? 0;
-
-                        plotMatched.influence_info = GetInfluenceInfoTotal(jsonContent.Value<JToken>("influence_info"), jsonContent.Value<Boolean?>("influence_poi_bonus") ?? false, posX, posY, jsonContent.Value<int?>("building_type_id") ?? 0);
-                        plotMatched.influence = jsonContent.Value<int?>("influence") ?? 0;
-                        plotMatched.influence_bonus = jsonContent.Value<int?>("influence_bonus") ?? 0;
-                        plotMatched.influence_poi_bonus = jsonContent.Value<Boolean?>("influence_poi_bonus") ?? false;
-                        plotMatched.production_poi_bonus = jsonContent.Value<decimal?>("production_poi_bonus") ?? 0.0m;
-                        plotMatched.is_perk_activated = jsonContent.Value<Boolean?>("is_perk_activated") ?? false;
-                        plotMatched.app_4_bonus = GetApplicationBonus(4, jsonContent.Value<JArray>("extra_appliances"), posX, posY);
-                        plotMatched.app_5_bonus = GetApplicationBonus(5, jsonContent.Value<JArray>("extra_appliances"), posX, posY);
-                        plotMatched.app_123_bonus = GetApplication123Bonus(jsonContent.Value<JToken>("appliances"), posX, posY);
-
-                        plotMatched.low_stamina_alert = citizen.CheckCitizenStamina(jsonContent.Value<JArray>("citizens"), plotMatched.building_type_id);
-                        plotMatched.action_id = jsonContent.Value<int?>("action_id") ?? 0;
-
-                        plotMatched.predict_produce = plotMatched.predict_produce == null ? 0 : plotMatched.predict_produce;
-
-                        returnPlot = plotMatched;
+                        _context.LogEvent(String.Concat("PlotDB:AddOrUpdatePlot() : Existing plot was found for X:", posX, " Y:", posY,". This maybe unexpected! Call was to create new plot at these XY coord. Existing Plot will be updated"));
                     }
                 }
+                    
+                if (plotId != 0 || plotMatched != null)
+                {
+                    plotMatched = plotMatched == null ? _context.plot.Find(plotId) : plotMatched;
+
+                    plotMatched.update_type = (int)UPDATE_TYPE.FULL;
+                    plotMatched.last_updated = DateTime.UtcNow;
+
+                    plotMatched.unclaimed_plot = string.IsNullOrEmpty(jsonContent.Value<string>("owner"));
+                    #nullable enable
+                    plotMatched.owner_nickname = jsonContent.Value<string?>("owner_nickname") ?? "";
+                    #nullable disable
+                    plotMatched.owner_matic = jsonContent.Value<string>("owner");
+                    plotMatched.owner_avatar_id = jsonContent.Value<int>("owner_avatar_id");
+                    plotMatched.resources = jsonContent.Value<int?>("resources") ?? 0;
+                    plotMatched.building_id = jsonContent.Value<int?>("building_id") ?? 0;
+                    plotMatched.building_level = jsonContent.Value<int?>("building_level") ?? 0;
+                    plotMatched.building_type_id = jsonContent.Value<int?>("building_type_id") ?? 0;
+                    plotMatched.token_id = jsonContent.Value<int?>("token_id") ?? 0;
+
+                    plotMatched.on_sale = jsonContent.Value<bool?>("on_sale") ?? false;
+                    plotMatched.current_price = plotMatched.on_sale ? building.GetSalePrice(jsonContent.Value<JToken>("sale_data"), worldType) : 0;
+                        
+                    plotMatched.for_rent = (jsonContent.Value<int?>("for_rent") ?? 0) > 0 ? building.GetRentPrice(jsonContent.Value<JToken>("rent_info"), worldType) : 0;
+                    plotMatched.rented = jsonContent.Value<string>("renter") != null;
+                    plotMatched.abundance = jsonContent.Value<int?>("abundance") ?? 0;
+                    plotMatched.building_abundance = plotMatched.building_abundance;        // only update from BuildingManage class calc
+                    plotMatched.condition = jsonContent.Value<int?>("condition") ?? 0;
+
+                    plotMatched.influence_info = GetInfluenceInfoTotal(jsonContent.Value<JToken>("influence_info"), jsonContent.Value<Boolean?>("influence_poi_bonus") ?? false, posX, posY, jsonContent.Value<int?>("building_type_id") ?? 0);
+                    
+                    newInfluence = jsonContent.Value<int?>("influence") ?? 0;
+                    plotMatched.influence_bonus = jsonContent.Value<int?>("influence_bonus") ?? 0;
+                    plotMatched.current_influence_rank = buildingManage.CheckInfluenceRankChange(newInfluence, plotMatched.influence ?? 0, plotMatched.influence_bonus ?? 0, plotMatched.current_influence_rank ?? 0, plotMatched.building_level, plotMatched.building_type_id);
+                    plotMatched.influence = newInfluence;                                   // Placed after ranking check, as both old and new influence needed for check
+                                        
+                    plotMatched.influence_poi_bonus = jsonContent.Value<Boolean?>("influence_poi_bonus") ?? false;
+                    plotMatched.production_poi_bonus = jsonContent.Value<decimal?>("production_poi_bonus") ?? 0.0m;
+                    plotMatched.is_perk_activated = jsonContent.Value<Boolean?>("is_perk_activated") ?? false;
+                    plotMatched.app_4_bonus = GetApplicationBonus(4, jsonContent.Value<JArray>("extra_appliances"), posX, posY);
+                    plotMatched.app_5_bonus = GetApplicationBonus(5, jsonContent.Value<JArray>("extra_appliances"), posX, posY);
+                    plotMatched.app_123_bonus = GetApplication123Bonus(jsonContent.Value<JToken>("appliances"), posX, posY);
+
+                    plotMatched.citizen_count = jsonContent.Value<JArray>("citizens") == null ? 0 : jsonContent.Value<JArray>("citizens").Count;
+                    plotMatched.low_stamina_alert = citizen.CheckCitizenStamina(jsonContent.Value<JArray>("citizens"), plotMatched.building_type_id);
+                    plotMatched.action_id = jsonContent.Value<int?>("action_id") ?? 0;
+
+                    plotMatched.predict_produce = plotMatched.predict_produce == null ? 0 : plotMatched.predict_produce;
+
+                    returnPlot = plotMatched;
+                }
+                
 
                 if (saveEvent)
                 {
@@ -331,7 +374,6 @@ namespace MetaverseMax.Database
         //      owner_avatar_id
         //      resources
         //      influence_info
-        //      influence_bonus
         //      influence_poi_bonus
         //      production_poi_bonus
         //      is_perk_activated
@@ -339,15 +381,19 @@ namespace MetaverseMax.Database
         //      app_5_bonus
         //      app_123_bonus
         //      action_id
-        //      for_rent
-        public bool UpdatePlotPartial(JToken ownerLand, bool saveEvent)
+        //      for_rent      >>> bool flag is not provided, BUT can infer if plot is currently rented by renter (matic) field and if not currently rented - if a rent price is assigned (0 =  not available for rent)
+        public PlotCord UpdatePlotPartial(JToken ownerLand, bool saveEvent)
         {
             Plot plotMatched = null;
             List<Plot> buildingPlotList = null;
             PlotManage plotManage = new(_context, worldType);
             CitizenManage citizen = new(_context, worldType);
+            BuildingManage buildingManage = new(_context, worldType);
             Building building = new();
-            int tokenId = 0;
+            int tokenId = 0, buildingLevel =0;
+            int newInfluence = 0; 
+            decimal newRanking = -1;
+            PlotCord plotFullUpdate = null;
 
             try
             {
@@ -358,72 +404,105 @@ namespace MetaverseMax.Database
                 if (buildingPlotList.Count == 0)
                 {
                     // Newly claimed plot in last 24 hrs - get(WS land/get) full plot details and update db.
-                    plotManage.AddOrUpdatePlot(0, ownerLand.Value<int?>("x") ?? 0, ownerLand.Value<int?>("y") ?? 0, saveEvent);
-                    logInfo(String.Concat("New plot token_id found - Getting full plot details and storing : tokenID - ", tokenId));
+                    plotMatched = plotManage.AddOrUpdatePlot(0, ownerLand.Value<int?>("x") ?? 0, ownerLand.Value<int?>("y") ?? 0, saveEvent);
+                    logInfo(String.Concat("New plot token_id found - Get full plot details and store : X: ", ownerLand.Value<int?>("x") ?? 0, "  Y:", ownerLand.Value<int?>("y") ?? 0));
                 }
                 else
                 {
                     plotMatched = buildingPlotList[0];
+                    buildingLevel = ownerLand.Value<int?>("building_level") ?? 0;
 
-                    // CHECK if owner changed == plot was transfered or sold since last sync, as owner_nickname and owner_avatar_id are not provided with thsi service call - need to get them
+                    // CHECK if owner changed == plot was transfered or sold since last sync, as owner_nickname and owner_avatar_id are not provided with this service call - need to get them
                     if (plotMatched.owner_matic.ToLower() != ownerLand.Value<string>("owner").ToLower())
                     {
                         // Newly claimed/transfer/sold plot in last 24 hrs - get(WS land/get) full plot details and update db.
                         // NOTE: This call will also update all related plots in building (MEGA / HUGE)
-                        plotManage.AddOrUpdatePlot(0, ownerLand.Value<int?>("x") ?? 0, ownerLand.Value<int?>("y") ?? 0, true);
-                        logInfo(String.Concat("Plot sold/transfered - Getting full plot details and storing - due to missing avatar & name: tokenID - ", tokenId));                        
+                        plotMatched = plotManage.AddOrUpdatePlot(plotMatched.plot_id, ownerLand.Value<int?>("x") ?? 0, ownerLand.Value<int?>("y") ?? 0, true);
+                        logInfo(String.Concat("Plot sold/transfer - Get full plot details and store - due to change of owner_matic, avatar, name.  Plot tokenID - ", tokenId));    
+                        
                     }
-
-                    else if (buildingPlotList.Count == 1 && ownerLand.Value<int?>("building_level") < (int)BUILDING_SIZE.HUGE)
+                    else if (buildingPlotList.Count == 1 && buildingLevel < (int)BUILDING_SIZE.HUGE)
                     {
                         // Update 1 plot or multiple plots depending on building level, only one record returned per building with this lands WS call.
                         // Notes: for_rent is not populated by this WS call, unknown if rental data is valid.  Dont change last_updated as not all fields are updated - specifically the influance_info field is not updated which used in ranking module to allow another refresh.
-
+                        plotMatched.update_type = (int)UPDATE_TYPE.PARTIAL;
                         plotMatched.last_updated = DateTime.UtcNow;
-                        plotMatched.current_price = building.GetSalePrice(ownerLand.Value<JToken>("sale_data"), worldType);
                         plotMatched.unclaimed_plot = string.IsNullOrEmpty(ownerLand.Value<string>("owner"));
-                        //plotMatched.owner_nickname = jsonContent.Value<string>("owner_nickname");
                         plotMatched.owner_matic = ownerLand.Value<string>("owner");
-                        //plotMatched.owner_avatar_id = jsonContent.Value<int>("owner_avatar_id");
-                        //plotMatched.resources = jsonContent.Value<int?>("resources") ?? 0;
+
                         plotMatched.building_id = ownerLand.Value<int?>("building_id") ?? 0;
-                        plotMatched.building_level = ownerLand.Value<int?>("building_level") ?? 0;
+                        plotMatched.building_level = buildingLevel;
                         plotMatched.building_type_id = ownerLand.Value<int?>("building_type_id") ?? 0;
                         plotMatched.token_id = tokenId;
                         plotMatched.on_sale = ownerLand.Value<bool?>("on_sale") ?? false;
                         plotMatched.current_price = plotMatched.on_sale ? building.GetSalePrice(ownerLand.Value<JToken>("sale_data"), worldType) : 0;
                         plotMatched.rented = ownerLand.Value<string>("renter") != null;
-                        plotMatched.for_rent = plotMatched.rented == false ? building.GetRentPrice(ownerLand.Value<JToken>("rent_info"), worldType) : 0;
+                        plotMatched.for_rent = plotMatched.rented == false ? building.GetRentPrice(ownerLand.Value<JToken>("rent_info"), worldType) : 0;    // Only get rent price if not currently rented
                         plotMatched.abundance = ownerLand.Value<int?>("abundance") ?? 0;
                         plotMatched.condition = ownerLand.Value<int?>("condition") ?? 0;
 
-                        plotMatched.influence = ownerLand.Value<int?>("influence") ?? 0;        // MCP calulated total IP with any POI & Monument included and ALWAYS active
+                        newInfluence = ownerLand.Value<int?>("influence") ?? 0;                                 // MCP calulated total IP with any POI & Monument included and ALWAYS active[using this WS]
+                        // Store plot data for later full update (if enabled)
+                        if (newInfluence != plotMatched.influence)
+                        {
+                            plotFullUpdate = new()
+                            {
+                                plotId = plotMatched.plot_id,
+                                posX = ownerLand.Value<int?>("x") ?? 0,
+                                posY = ownerLand.Value<int?>("y") ?? 0
+                            };                            
+                        }
+                        plotMatched.current_influence_rank = buildingManage.CheckInfluenceRankChange(newInfluence, plotMatched.influence ?? 0, plotMatched.influence_bonus ?? 0, plotMatched.current_influence_rank ?? 0, buildingLevel, plotMatched.building_type_id);
+                        plotMatched.influence = newInfluence;                                                   // Placed after ranking check, as both old and new influence needed for check                    
+                        
+                        //plotMatched.influence_bonus = ownerLand.Value<int?>("influence_bonus") ?? 0;          // Missing assign bonus per app slot
 
+                        plotMatched.citizen_count = ownerLand.Value<JArray>("citizens") == null ? 0 : ownerLand.Value<JArray>("citizens").Count;
                         plotMatched.low_stamina_alert = citizen.CheckCitizenStamina(ownerLand.Value<JArray>("citizens"), plotMatched.building_type_id);
                     }
-                    else if (ownerLand.Value<int?>("building_level") == (int)BUILDING_SIZE.HUGE || ownerLand.Value<int?>("building_level") == (int)BUILDING_SIZE.MEGA)
+                    else if ((ownerLand.Value<int?>("building_level") == (int)BUILDING_SIZE.HUGE || ownerLand.Value<int?>("building_level") == (int)BUILDING_SIZE.MEGA) || buildingPlotList.Count > 1)
                     {
+                        // On Influence change - tag token for full update.
+                        newInfluence = ownerLand.Value<int?>("influence") ?? 0;
+                        if (buildingPlotList.Count > 0 && newInfluence != buildingPlotList[0].influence)
+                        {
+                            plotFullUpdate = new()
+                            {
+                                plotId = plotMatched.plot_id,
+                                posX = ownerLand.Value<int?>("x") ?? 0,
+                                posY = ownerLand.Value<int?>("y") ?? 0
+                            };
+                        }
 
                         for (int i = 0; i < buildingPlotList.Count; i++)
                         {
                             plotMatched = buildingPlotList[i];
 
+                            plotMatched.update_type = (int)UPDATE_TYPE.PARTIAL;
                             plotMatched.last_updated = DateTime.UtcNow;
-                            plotMatched.current_price = building.GetSalePrice(ownerLand.Value<JToken>("sale_data"), worldType);
                             plotMatched.unclaimed_plot = string.IsNullOrEmpty(ownerLand.Value<string>("owner"));
                             plotMatched.owner_matic = ownerLand.Value<string>("owner");
+
                             plotMatched.building_id = ownerLand.Value<int?>("building_id") ?? 0;
-                            plotMatched.building_level = ownerLand.Value<int?>("building_level") ?? 0;
+                            plotMatched.building_level = buildingLevel;
                             plotMatched.building_type_id = ownerLand.Value<int?>("building_type_id") ?? 0;
                             plotMatched.token_id = ownerLand.Value<int?>("token_id") ?? 0;
                             plotMatched.on_sale = ownerLand.Value<bool?>("on_sale") ?? false;
                             plotMatched.current_price = plotMatched.on_sale ? building.GetSalePrice(ownerLand.Value<JToken>("sale_data"), worldType) : 0;
                             plotMatched.rented = ownerLand.Value<string>("renter") != null;
                             plotMatched.for_rent = plotMatched.rented == false ? building.GetRentPrice(ownerLand.Value<JToken>("rent_info"), worldType) : 0;
-                            //plotMatched.abundance = ownerLand.Value<int?>("abundance") ?? 0;          // dont update abundance for related building plots - each plot can have own abundance
+                            //plotMatched.abundance = ownerLand.Value<int?>("abundance") ?? 0;                  // dont update abundance for related building plots - each plot can have own abundance
                             plotMatched.condition = ownerLand.Value<int?>("condition") ?? 0;
+                                                      
+                            if (newRanking == -1) {
+                                newRanking = buildingManage.CheckInfluenceRankChange(newInfluence, plotMatched.influence ?? 0, plotMatched.influence_bonus ?? 0, plotMatched.current_influence_rank ?? 0, buildingLevel, plotMatched.building_type_id);
+                            }
+                            plotMatched.current_influence_rank = newRanking;                                    // Reuse ranking if already identified on prior related plot
+                            plotMatched.influence = newInfluence;                                               // Placed after ranking check, as both old and new influence needed for check
 
-                            plotMatched.influence = ownerLand.Value<int?>("influence") ?? 0;
+                            //plotMatched.influence_bonus = ownerLand.Value<int?>("influence_bonus") ?? 0;      // MCP calulated total IP with any POI & Monument included and ALWAYS active
+
+                            plotMatched.citizen_count = ownerLand.Value<JArray>("citizens") == null ? 0 : ownerLand.Value<JArray>("citizens").Count;
                             plotMatched.low_stamina_alert = citizen.CheckCitizenStamina(ownerLand.Value<JArray>("citizens"), plotMatched.building_type_id);
                         }
                     }
@@ -439,7 +518,7 @@ namespace MetaverseMax.Database
                 logException(ex, String.Concat("PlotDB:UpdatePlotPartial() : Error Adding/update Plot token_id: ", tokenId));
             }
 
-            return true;
+            return plotFullUpdate;
         }
 
         private int GetApplicationBonus(int appNumber, JArray extraAppliances, int pos_x, int pos_y)
@@ -464,17 +543,7 @@ namespace MetaverseMax.Database
                     }
 
                     if (appType > 0) {
-                        appBonus = appType switch
-                        {
-                            7 => 10,    // Red Sat (7) = 10% bonus
-                            5 => 8,     // White Sat
-                            8 => 6,     // Green Air con
-                            1 => 5,     // White Air con
-                            2 => 3,     // CCV White                            
-                            4 => 2,     // Router Black 
-                            3 => 1,     // Red Fire Alarm
-                            _ => 0
-                        };
+                        appBonus = (int)FindApplicationBonusValue((APPLICATION_ID)appType);                        
                     }
                 }
             }
@@ -488,7 +557,8 @@ namespace MetaverseMax.Database
 
         private int GetApplication123Bonus(JToken Appliance123, int pos_x, int pos_y)
         {
-            int appBonus = 0, appType = 0;
+            int appBonusTotal = 0;
+
             try
             {
                 if (Appliance123 != null && Appliance123.Any())
@@ -497,22 +567,11 @@ namespace MetaverseMax.Database
 
                     for (int appCount = 0; appCount < values.Count; appCount++)
                     {
-                        appType = values[appCount] ?? 0;
+                        int appType = values[appCount] ?? 0;
 
                         if (appType > 0)
-                        {
-                            appBonus += appType switch
-                            {
-                                (int)APPLICATION.RED_SAT => 10,    // Red Sat (7) = 10% bonus
-                                (int)APPLICATION.WHITE_SAT => 8,     // White Sat
-                                (int)APPLICATION.GREEN_AIR_CON => 6,     // Green Air con
-                                (int)APPLICATION.WHITE_AIR_CON => 5,     // White Air con
-                                (int)APPLICATION.CCTV_RED => 4,     // CCTV RED 
-                                (int)APPLICATION.CCTV_WHITE => 3,     // CCTV White                            
-                                (int)APPLICATION.ROUTER_BLACK => 2,     // Router Black 
-                                (int)APPLICATION.RED_FIRE_ALARM => 1,     // Red Fire Alarm
-                                _ => 0
-                            };
+                        {                            
+                            appBonusTotal += (int)FindApplicationBonusValue((APPLICATION_ID)appType);
                         }
                     }
                 }
@@ -521,6 +580,27 @@ namespace MetaverseMax.Database
             {
                 logException(ex, String.Concat("PlotDB.GetApplication123Bonus() : Error calculating app bonus info for plot X:", pos_x, " Y:", pos_y));
             }
+
+            return appBonusTotal;
+        }
+
+        APPLICATION_BONUS FindApplicationBonusValue(APPLICATION_ID appType)
+        {
+            APPLICATION_BONUS appBonus;
+            appBonus = appType switch
+            {
+                APPLICATION_ID.RED_SAT => APPLICATION_BONUS.RED_SAT,                
+                APPLICATION_ID.GREEN_SAT => APPLICATION_BONUS.GREEN_SAT,
+                APPLICATION_ID.WHITE_SAT => APPLICATION_BONUS.WHITE_SAT,
+                APPLICATION_ID.GREEN_AIR_CON => APPLICATION_BONUS.GREEN_AIR_CON,
+                APPLICATION_ID.WHITE_AIR_CON => APPLICATION_BONUS.WHITE_AIR_CON,
+                APPLICATION_ID.RED_AIR_CON => APPLICATION_BONUS.RED_AIR_CON,
+                APPLICATION_ID.CCTV_RED => APPLICATION_BONUS.CCTV_RED,
+                APPLICATION_ID.CCTV_WHITE => APPLICATION_BONUS.CCTV_WHITE,
+                APPLICATION_ID.ROUTER_BLACK => APPLICATION_BONUS.ROUTER_BLACK,
+                APPLICATION_ID.RED_FIRE_ALARM => APPLICATION_BONUS.RED_FIRE_ALARM,
+                _ => 0
+            };
 
             return appBonus;
         }
