@@ -86,7 +86,7 @@ namespace MetaverseMax.ServiceClass
             return districtPlots;
         }
 
-        public string BuildingImg(int buildingType, int buildingID, int buildingLvl)
+        public string GetBuildingImg(int buildingType, int buildingID, int buildingLvl, WORLD_TYPE worldType)
         {
             string buildingImg = string.Empty;
 
@@ -95,7 +95,7 @@ namespace MetaverseMax.ServiceClass
                 case (int)BUILDING_TYPE.RESIDENTIAL:
                     buildingImg = buildingID switch
                     {
-                        (int)BUILDING_SUBTYPE.COMDOMINIUM => buildingLvl switch
+                        (int)BUILDING_SUBTYPE.CONDOMINIUM => buildingLvl switch
                         {
                             <= 5 => "https://play.mcp3d.com/assets/images/buildings/Residential3_LVL" + Convert.ToString(buildingLvl) + "-1.png",
                             6 => "https://play.mcp3d.com/assets/images/buildings/Residential3_Huge-1.png",
@@ -124,7 +124,7 @@ namespace MetaverseMax.ServiceClass
                             _ => "https://play.mcp3d.com/assets/images/buildings/Residential3_LVL1-1.png"
                         },
                     };
-                    break; //buildingImg = "29";
+                    break;
 
                 case (int)BUILDING_TYPE.ENERGY:
                     buildingImg = buildingID switch
@@ -151,7 +151,7 @@ namespace MetaverseMax.ServiceClass
                             _ => "https://play.mcp3d.com/assets/images/buildings/Energy_ThermalPower_LVL1-1.png"
                         },
                     };                                                          
-                    break;//buildingImg = "5"; 
+                    break;
 
                 case (int)BUILDING_TYPE.COMMERCIAL:
                     buildingImg = buildingID switch
@@ -178,7 +178,7 @@ namespace MetaverseMax.ServiceClass
                             _ => "https://play.mcp3d.com/assets/images/buildings/Blockmart_V2_LVL1-1.png"
                         },
                     };                    
-                    break; //buildingImg = "64";
+                    break;
 
                 case (int)BUILDING_TYPE.INDUSTRIAL:
                     buildingImg = buildingID switch
@@ -232,7 +232,7 @@ namespace MetaverseMax.ServiceClass
                             _ => "https://play.mcp3d.com/assets/images/buildings/OfficeRing_LVL1-1.png"
                         },
                     };
-                    break;// buildingImg = "30";
+                    break;
                     
                 case (int)BUILDING_TYPE.PRODUCTION:
                     buildingImg = buildingID switch
@@ -251,19 +251,57 @@ namespace MetaverseMax.ServiceClass
                             7 => "https://play.mcp3d.com/assets/images/buildings/BrickFactory_V2_Mega-1.png",
                             _ => "https://play.mcp3d.com/assets/images/buildings/BrickFactory_V2_LVL1-1.png"
                         },
-                        (int)BUILDING_SUBTYPE.CONCRETE_PLANT => buildingLvl switch
+                        (int)BUILDING_SUBTYPE.CONCRETE_PLANT => worldType switch
                         {
-                            <= 5 => "https://play.mcp3d.com/assets/images/buildings/ConcreteMill_LVL" + Convert.ToString(buildingLvl) + "-1.png",
-                            6 => "https://play.mcp3d.com/assets/images/buildings/ConcreteMill_Huge-1.png",
-                            7 => "https://play.mcp3d.com/assets/images/buildings/ConcreteMill_Mega-1.png",
-                            _ => "https://play.mcp3d.com/assets/images/buildings/ConcreteMill_LVL1-1.png"
-                        },
+                            WORLD_TYPE.TRON => buildingLvl switch
+                            {
+                                <= 5 => "https://play.mcp3d.com/assets/images/buildings/ConcreteMill_LVL" + Convert.ToString(buildingLvl) + "-1.png",
+                                6 => "https://play.mcp3d.com/assets/images/buildings/ConcreteMill_Huge-1.png",
+                                7 => "https://play.mcp3d.com/assets/images/buildings/ConcreteMill_Mega-1.png",
+                                _ => "https://play.mcp3d.com/assets/images/buildings/ConcreteMill_LVL1-1.png"
+                            },
+                            WORLD_TYPE.ETH => buildingLvl switch
+                            {
+                                <= 5 => "https://play.mcp3d.com/assets/images/buildings/SteelMill_V2_LVL" + Convert.ToString(buildingLvl) + "-1.png",
+                                6 => "https://play.mcp3d.com/assets/images/buildings/SteelMill_V2_Huge-1.png",
+                                7 => "https://play.mcp3d.com/assets/images/buildings/SteelMill_V2_Mega-1.png",
+                                _ => "https://play.mcp3d.com/assets/images/buildings/SteelMill_V2_LVL1-1.png"
+                            },
+                            WORLD_TYPE.BNB => buildingLvl switch
+                            {
+                                <= 5 => "https://play.mcp3d.com/assets/images/buildings/Plastic_LVL" + Convert.ToString(buildingLvl) + "-1.png",
+                                6 => "https://play.mcp3d.com/assets/images/buildings/Plastic_Huge-1.png",
+                                7 => "https://play.mcp3d.com/assets/images/buildings/Plastic_Mega-1.png",
+                                _ => "https://play.mcp3d.com/assets/images/buildings/Plastic_LVL1-1.png"
+                            },
+                            _ => buildingLvl switch
+                            {
+                                <= 5 => "https://play.mcp3d.com/assets/images/buildings/ConcreteMill_LVL" + Convert.ToString(buildingLvl) + "-1.png",
+                                6 => "https://play.mcp3d.com/assets/images/buildings/ConcreteMill_Huge-1.png",
+                                7 => "https://play.mcp3d.com/assets/images/buildings/ConcreteMill_Mega-1.png",
+                                _ => "https://play.mcp3d.com/assets/images/buildings/ConcreteMill_LVL1-1.png"
+                            },
+                        },                      
                         (int)BUILDING_SUBTYPE.FACTORY => buildingLvl switch
                         {
                             <= 5 => "https://play.mcp3d.com/assets/images/buildings/Factory_V2_LVL" + Convert.ToString(buildingLvl) + "-1.png",
-                            6 => "https://play.mcp3d.com/assets/images/buildings//Factory_V2_Huge-1.png",
+                            6 => "https://play.mcp3d.com/assets/images/buildings/Factory_V2_Huge-1.png",
                             7 => "https://play.mcp3d.com/assets/images/buildings/Factory_V2_Mega-1.png",
                             _ => "https://play.mcp3d.com/assets/images/buildings/Factory_V2_LVL1-1.png"
+                        },
+                        (int)BUILDING_SUBTYPE.CHEMICAL_PLANT => buildingLvl switch
+                        {
+                            <= 5 => "https://play.mcp3d.com/assets/images/buildings/ChemicalPlant_LVL" + Convert.ToString(buildingLvl) + "-1.png",
+                            6 => "https://play.mcp3d.com/assets/images/buildings/ChemicalPlant_Huge-1.png",
+                            7 => "https://play.mcp3d.com/assets/images/buildings/ChemicalPlant_Mega-1.png",
+                            _ => "https://play.mcp3d.com/assets/images/buildings/ChemicalPlant_LVL1-1.png"
+                        },
+                        (int)BUILDING_SUBTYPE.PAPER_FACTORY => buildingLvl switch
+                        {
+                            <= 5 => "https://play.mcp3d.com/assets/images/buildings/PaperFactory_LVL" + Convert.ToString(buildingLvl) + "-1.png",
+                            6 => "https://play.mcp3d.com/assets/images/buildings/PaperFactory_Huge-1.png",
+                            7 => "https://play.mcp3d.com/assets/images/buildings/PaperFactory_Mega-1.png",
+                            _ => "https://play.mcp3d.com/assets/images/buildings/PaperFactory_LVL1-1.png"
                         },
                         _ => buildingLvl switch
                         {
@@ -306,18 +344,18 @@ namespace MetaverseMax.ServiceClass
                             7 => "https://play.mcp3d.com/assets/images/buildings/Hospital_V2_Mega-1.png",
                             _ => "https://play.mcp3d.com/assets/images/buildings/Hospital_V2_LVL1-1.png"
                         },
-                    }; // buildingImg = "28";
+                    };
                     break;
 
                 case (int)BUILDING_TYPE.POI:                    
                     buildingImg = buildingID switch
                     {
-                        100 => "./images/OfficeMonument.PNG",
-                        104 => "./images/IndustrialMonument.PNG",
-                        106 => "./images/ProductionMonument.PNG",
-                        108 => "./images/CommercialMonument.PNG",
-                        110 => "./images/ResidentialMonument.PNG",
-                        112 => "./images/EnergyMonument.PNG",
+                        (int)BUILDING_SUBTYPE.OFFICE_MONUMENT => "./images/OfficeMonument.PNG",
+                        (int)BUILDING_SUBTYPE.INDUSTRY_MONUMENT => "./images/IndustrialMonument.PNG",
+                        (int)BUILDING_SUBTYPE.PRODUCTION_MONUMENT => "./images/ProductionMonument.PNG",
+                        (int)BUILDING_SUBTYPE.COMMERCIAL_MONUMENT => "./images/CommercialMonument.PNG",
+                        (int)BUILDING_SUBTYPE.RESIDENTIAL_MONUMENT => "./images/ResidentialMonument.PNG",
+                        (int)BUILDING_SUBTYPE.ENERGY_MONUMENT => "./images/EnergyMonument.PNG",
                         _ => "./images/POI.png",
                     };
                     break;
